Application finished.

+++ app.js
@@ -3,7 +3,7 @@ require('colors');
 // ================================================================================ //
 
 // =============================== My imports ===================================== //
-const { menu, pause, readInput, listTasksForDelete, confirmation } = require('./helpers/inquirer');
+const { menu, pause, readInput, listTasksForDelete, confirmation, showCheckList } = require('./helpers/inquirer');
 const { saveData, loadData } = require('./helpers/save-data');
 const Tasks = require('./models/tasks');
 // =============================================================================== //
@@ -25,35 +25,50 @@ const Tasks = require('./models/tasks');
         selectedOption = await menu();
 
         switch (selectedOption) {
+
             case 1:
                 const taskDescription = await readInput('Write task description:');
                 tasks.addTask(taskDescription);
                 break;
+
             case 2:
                 tasks.listTasksWithFormat();
                 break;
+
             case 3:
                 tasks.listFinishedAndNotFinished();
                 break;
+
             case 4:
                 tasks.listFinishedAndNotFinished(false);
-                break
+                break;
+
+            case 5:
+                const ids = await showCheckList(tasks.tasksArr);
+                tasks.toggleCompletedTasks(ids);
+                break;
+
             case 6:
 
                 const id = await listTasksForDelete(tasks.tasksArr);
-                const ok = await confirmation(`Are you sure?`);
 
-                if (ok) {
+                if (id !== 0) {
 
-                    let temporalDescription = tasks.list[id].description;
-                    tasks.deleteTask(id);
-                    console.log('\nThe task \"' + `${temporalDescription}`.blue + '\" was deleted.');
+                    const ok = await confirmation(`Are you sure?`);
 
+                    if (ok) {
+
+
+                        let temporalDescription = tasks.list[id].description;
+                        tasks.deleteTask(id);
+                        console.log('\nThe task \"' + `${temporalDescription}`.blue + '\" was deleted.');
+
+                    }
                 }
+
 
                 break;
             default:
-                await pause();
                 break;
         }
 
@@ -62,5 +77,7 @@ const Tasks = require('./models/tasks');
         saveData(tasks.tasksArr);
 
     } while (selectedOption !== 7);
+
+    console.clear();
 
 })();
